resin-proxy-config: Add recipe for service to configure and start redsocks
We add a resin-proxy-config service that starts a script. This script looks for
a system-proxy folder in the boot partition that should have a [redsocks.conf file](https://github.com/darkk/redsocks/blob/master/redsocks.conf.example).

If this file is present, it copies it to /etc/redsocks/ and starts the redsocks program, creating iptables rules
to redirect all tcp traffic to port 12345 except local and docker connections.

This means that the redsocks.conf has to configure redsocks to listen on port 12345 (with the local_port option) and set `daemon = off` to work properly.

Signed-off-by: Pablo Carranza Velez <pablo@resin.io>